test: align tests with integration modules
