+++ tests/test_init.py
@@ -1,45 +1,31 @@
-"""End-to-end config-entry setup tests for the Variable integration."""
+"""End-to-end setup orchestration tests for the Variable integration."""
 
 from collections.abc import Callable, Mapping
 from typing import Any
 from unittest.mock import AsyncMock, patch
 
-from homeassistant.components.device_tracker.const import ATTR_LOCATION_NAME
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import (
     ATTR_BATTERY_LEVEL,
-    ATTR_CONFIGURATION_URL,
     ATTR_GPS_ACCURACY,
     ATTR_LATITUDE,
     ATTR_LONGITUDE,
-    ATTR_MANUFACTURER,
-    ATTR_MODEL,
-    CONF_DEVICE,
-    CONF_NAME,
     SERVICE_RELOAD,
     STATE_ON,
     Platform,
 )
 from homeassistant.core import HomeAssistant
-from homeassistant.data_entry_flow import FlowResultType
-from homeassistant.helpers import device_registry as dr, entity_registry as er
+from homeassistant.helpers import entity_registry as er
 from homeassistant.setup import async_setup_component
 import pytest
 
 from custom_components.variable.const import (
-    ATTR_ATTRIBUTES,
-    ATTR_REPLACE_ATTRIBUTES,
-    CONF_ATTRIBUTES,
     CONF_ENTITY_PLATFORM,
-    CONF_EXCLUDE_FROM_RECORDER,
-    CONF_FORCE_UPDATE,
-    CONF_RESTORE,
     CONF_VALUE,
     CONF_VARIABLE_ID,
     CONF_YAML_PRESENT,
     CONF_YAML_VARIABLE,
     DOMAIN,
-    SERVICE_UPDATE_SENSOR,
 )
 
 ConfigEntryFactory = Callable[[Mapping[str, Any]], ConfigEntry]
@@ -112,30 +98,6 @@ async def test_yaml_setup_and_reload_manage_config_entries(
     assert hass.config_entries.async_get_entry(removed_entry.entry_id) is None
     assert hass.states.get("sensor.yaml_removed") is None
     assert er.async_get(hass).async_get("sensor.yaml_removed") is None
-
-
-async def test_yaml_entry_aborts_options_flow(hass: HomeAssistant) -> None:
-    """Reject options for a variable managed through YAML.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-    """
-    assert await async_setup_component(
-        hass,
-        DOMAIN,
-        {DOMAIN: {"yaml_managed": {CONF_VALUE: "managed"}}},
-    )
-    await hass.async_block_till_done()
-    entry = next(
-        entry
-        for entry in hass.config_entries.async_entries(DOMAIN)
-        if entry.data.get(CONF_VARIABLE_ID) == "yaml_managed"
-    )
-
-    result = await hass.config_entries.options.async_init(entry.entry_id)
-
-    assert result["type"] is FlowResultType.ABORT
-    assert result["reason"] == "yaml_variable"
 
 
 @pytest.mark.parametrize(
@@ -222,36 +184,6 @@ async def test_setup_entry_creates_platform_entity(
     assert registry_entry.config_entry_id == entry.entry_id
 
 
-async def test_setup_device_entry_creates_registry_device(
-    hass: HomeAssistant,
-    config_entry_factory: ConfigEntryFactory,
-) -> None:
-    """Load a device config entry and register its metadata.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-        config_entry_factory: Factory that creates the device config entry.
-    """
-    entry = config_entry_factory(
-        {
-            CONF_ENTITY_PLATFORM: CONF_DEVICE,
-            CONF_NAME: "Virtual Hub",
-            CONF_YAML_VARIABLE: False,
-            ATTR_MANUFACTURER: "Variables",
-            ATTR_CONFIGURATION_URL: "https://example.com/device",
-        }
-    )
-
-    assert await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-
-    device = dr.async_get(hass).async_get_device(identifiers={(DOMAIN, entry.entry_id)})
-    assert device is not None
-    assert device.name == "Virtual Hub"
-    assert device.manufacturer == "Variables"
-    assert str(device.configuration_url) == "https://example.com/device"
-
-
 async def test_unload_entry_removes_entity(
     hass: HomeAssistant,
     sensor_entry: ConfigEntry,
@@ -271,443 +203,3 @@ async def test_unload_entry_removes_entity(
 
     assert hass.states.get("sensor.office_temperature") is None
     assert er.async_get(hass).async_get("sensor.office_temperature") is None
-
-
-async def test_options_flow_changes_loaded_sensor_value(
-    hass: HomeAssistant,
-    sensor_entry: ConfigEntry,
-) -> None:
-    """Change a live entity through the options workflow and its entity service.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-        sensor_entry: Loaded sensor config entry to update.
-    """
-    assert await hass.config_entries.async_setup(sensor_entry.entry_id)
-    await hass.async_block_till_done()
-
-    result = await hass.config_entries.options.async_init(sensor_entry.entry_id)
-    assert result["type"] is FlowResultType.MENU
-    assert result["menu_options"] == ["change_sensor_value", "sensor_options"]
-
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={"next_step_id": "change_sensor_value"},
-    )
-    assert result["type"] is FlowResultType.FORM
-
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={CONF_VALUE: "24.25", "attributes": {"source": "options"}},
-    )
-    await hass.async_block_till_done()
-
-    assert result["type"] is FlowResultType.ABORT
-    assert result["reason"] == "value_changed"
-    state = hass.states.get("sensor.office_temperature")
-    assert state is not None
-    assert state.state == "24.25"
-    assert state.attributes["source"] == "options"
-
-
-async def test_sensor_options_flow_updates_entry_and_live_entity(
-    hass: HomeAssistant,
-    sensor_entry: ConfigEntry,
-) -> None:
-    """Persist both sensor option pages and reload the live entity.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-        sensor_entry: Loaded sensor config entry to update.
-    """
-    assert await hass.config_entries.async_setup(sensor_entry.entry_id)
-    await hass.async_block_till_done()
-
-    result = await hass.config_entries.options.async_init(sensor_entry.entry_id)
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={"next_step_id": "sensor_options"},
-    )
-    assert result["type"] is FlowResultType.FORM
-    assert result["step_id"] == "sensor_options"
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={
-            "device_class": "None",
-            CONF_RESTORE: False,
-            CONF_FORCE_UPDATE: True,
-            CONF_EXCLUDE_FROM_RECORDER: True,
-        },
-    )
-    assert result["type"] is FlowResultType.FORM
-    assert result["step_id"] == "sensor_options_page_2"
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={
-            CONF_VALUE: "26.5",
-            CONF_ATTRIBUTES: {"source": "sensor-options"},
-        },
-    )
-    await hass.async_block_till_done()
-
-    assert result["type"] is FlowResultType.CREATE_ENTRY
-    assert sensor_entry.data[CONF_VALUE] == "26.5"
-    assert sensor_entry.data[CONF_FORCE_UPDATE] is True
-    assert sensor_entry.data[CONF_EXCLUDE_FROM_RECORDER] is True
-    assert sensor_entry.data[CONF_ATTRIBUTES] == {"source": "sensor-options"}
-    state = hass.states.get("sensor.office_temperature")
-    assert state is not None
-    assert state.state == "26.5"
-    assert state.attributes["source"] == "sensor-options"
-
-
-async def test_binary_sensor_options_update_entry_and_live_entity(
-    hass: HomeAssistant,
-    config_entry_factory: ConfigEntryFactory,
-) -> None:
-    """Persist binary sensor options and reload its state and attributes.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-        config_entry_factory: Factory that creates the binary-sensor config entry.
-    """
-    entry = config_entry_factory(
-        {
-            CONF_ENTITY_PLATFORM: Platform.BINARY_SENSOR,
-            CONF_VARIABLE_ID: "configured_binary",
-            CONF_NAME: "Configured Binary",
-            CONF_VALUE: "true",
-            CONF_YAML_VARIABLE: False,
-            CONF_RESTORE: False,
-            CONF_FORCE_UPDATE: False,
-        }
-    )
-    assert await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-
-    result = await hass.config_entries.options.async_init(entry.entry_id)
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={"next_step_id": "binary_sensor_options"},
-    )
-    assert result["type"] is FlowResultType.FORM
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={
-            CONF_VALUE: "false",
-            CONF_ATTRIBUTES: {"source": "binary-options-form"},
-            "device_class": "door",
-            CONF_RESTORE: False,
-            CONF_FORCE_UPDATE: True,
-            CONF_EXCLUDE_FROM_RECORDER: False,
-        },
-    )
-    await hass.async_block_till_done()
-
-    assert result["type"] is FlowResultType.CREATE_ENTRY
-    assert entry.data[CONF_VALUE] == "false"
-    assert entry.data["device_class"] == "door"
-    assert entry.data[CONF_ATTRIBUTES] == {"source": "binary-options-form"}
-    assert entry.data[CONF_RESTORE] is False
-    assert entry.data[CONF_FORCE_UPDATE] is True
-    assert entry.data[CONF_EXCLUDE_FROM_RECORDER] is False
-    state = hass.states.get("binary_sensor.configured_binary")
-    assert state is not None
-    assert state.state == "off"
-    assert state.attributes["source"] == "binary-options-form"
-
-
-async def test_device_tracker_options_update_entry_and_live_entity(
-    hass: HomeAssistant,
-    config_entry_factory: ConfigEntryFactory,
-) -> None:
-    """Persist tracker options and reload its coordinates and attributes.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-        config_entry_factory: Factory that creates the device-tracker config entry.
-    """
-    entry = config_entry_factory(
-        {
-            CONF_ENTITY_PLATFORM: Platform.DEVICE_TRACKER,
-            CONF_VARIABLE_ID: "configured_tracker",
-            CONF_NAME: "Configured Tracker",
-            CONF_YAML_VARIABLE: False,
-            ATTR_LATITUDE: 40.0,
-            ATTR_LONGITUDE: -75.0,
-            ATTR_GPS_ACCURACY: 10,
-            ATTR_BATTERY_LEVEL: 80,
-            CONF_RESTORE: False,
-            CONF_FORCE_UPDATE: False,
-        }
-    )
-    assert await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-
-    result = await hass.config_entries.options.async_init(entry.entry_id)
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={"next_step_id": "device_tracker_options"},
-    )
-    assert result["type"] is FlowResultType.FORM
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={
-            ATTR_LATITUDE: 41.25,
-            ATTR_LONGITUDE: -74.75,
-            ATTR_LOCATION_NAME: "Studio",
-            ATTR_GPS_ACCURACY: 4,
-            ATTR_BATTERY_LEVEL: 65,
-            CONF_ATTRIBUTES: {"source": "tracker-options-form"},
-            CONF_RESTORE: False,
-            CONF_FORCE_UPDATE: True,
-            CONF_EXCLUDE_FROM_RECORDER: False,
-        },
-    )
-    await hass.async_block_till_done()
-
-    assert result["type"] is FlowResultType.CREATE_ENTRY
-    assert entry.data[ATTR_LATITUDE] == 41.25
-    assert entry.data[ATTR_LONGITUDE] == -74.75
-    assert entry.data[ATTR_LOCATION_NAME] == "Studio"
-    assert entry.data[CONF_ATTRIBUTES] == {"source": "tracker-options-form"}
-    assert entry.data[CONF_RESTORE] is False
-    assert entry.data[CONF_FORCE_UPDATE] is True
-    assert entry.data[CONF_EXCLUDE_FROM_RECORDER] is False
-    state = hass.states.get("device_tracker.configured_tracker")
-    assert state is not None
-    assert state.state == "Studio"
-    assert state.attributes[ATTR_LATITUDE] == 41.25
-    assert state.attributes[ATTR_LONGITUDE] == -74.75
-    assert state.attributes[ATTR_GPS_ACCURACY] == 4
-    assert state.attributes[ATTR_BATTERY_LEVEL] == 65
-    assert state.attributes["source"] == "tracker-options-form"
-
-
-async def test_sensor_entity_service_updates_state_and_attributes(
-    hass: HomeAssistant,
-    sensor_entry: ConfigEntry,
-) -> None:
-    """Mutate a loaded sensor through Home Assistant's registered entity service.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-        sensor_entry: Loaded sensor config entry to update.
-    """
-    assert await hass.config_entries.async_setup(sensor_entry.entry_id)
-    await hass.async_block_till_done()
-
-    await hass.services.async_call(
-        DOMAIN,
-        SERVICE_UPDATE_SENSOR,
-        {
-            "entity_id": ["sensor.office_temperature"],
-            CONF_VALUE: 19,
-            ATTR_ATTRIBUTES: {"service_marker": True},
-            ATTR_REPLACE_ATTRIBUTES: True,
-        },
-        blocking=True,
-    )
-
-    state = hass.states.get("sensor.office_temperature")
-    assert state is not None
-    assert state.state == "19"
-    assert state.attributes["service_marker"] is True
-    assert "source" not in state.attributes
-
-
-async def test_binary_sensor_options_flow_changes_live_entity(
-    hass: HomeAssistant,
-    config_entry_factory: ConfigEntryFactory,
-) -> None:
-    """Change binary state and attributes through the options workflow.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-        config_entry_factory: Factory that creates the binary-sensor config entry.
-    """
-    entry = config_entry_factory(
-        {
-            CONF_ENTITY_PLATFORM: Platform.BINARY_SENSOR,
-            CONF_VARIABLE_ID: "options_binary",
-            CONF_NAME: "Options Binary",
-            CONF_VALUE: "true",
-            CONF_YAML_VARIABLE: False,
-            "restore": False,
-            "force_update": False,
-        }
-    )
-    assert await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-
-    result = await hass.config_entries.options.async_init(entry.entry_id)
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={"next_step_id": "change_binary_sensor_value"},
-    )
-    assert result["type"] is FlowResultType.FORM
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={CONF_VALUE: "false", "attributes": {"source": "binary-options"}},
-    )
-    await hass.async_block_till_done()
-
-    assert result["type"] is FlowResultType.ABORT
-    assert result["reason"] == "value_changed"
-    state = hass.states.get("binary_sensor.options_binary")
-    assert state is not None
-    assert state.state == "off"
-    assert state.attributes["source"] == "binary-options"
-
-
-async def test_device_tracker_options_flow_changes_live_entity(
-    hass: HomeAssistant,
-    config_entry_factory: ConfigEntryFactory,
-) -> None:
-    """Change tracker coordinates and metadata through the options workflow.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-        config_entry_factory: Factory that creates the device-tracker config entry.
-    """
-    entry = config_entry_factory(
-        {
-            CONF_ENTITY_PLATFORM: Platform.DEVICE_TRACKER,
-            CONF_VARIABLE_ID: "options_tracker",
-            CONF_NAME: "Options Tracker",
-            CONF_YAML_VARIABLE: False,
-            ATTR_LATITUDE: 40.0,
-            ATTR_LONGITUDE: -75.0,
-            ATTR_GPS_ACCURACY: 10,
-            ATTR_BATTERY_LEVEL: 80,
-            "restore": False,
-            "force_update": False,
-        }
-    )
-    assert await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-
-    result = await hass.config_entries.options.async_init(entry.entry_id)
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={"next_step_id": "change_device_tracker_value"},
-    )
-    assert result["type"] is FlowResultType.FORM
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={
-            ATTR_LATITUDE: 41.5,
-            ATTR_LONGITUDE: -74.5,
-            ATTR_GPS_ACCURACY: 6,
-            ATTR_BATTERY_LEVEL: 72,
-            "attributes": {"source": "tracker-options"},
-        },
-    )
-    await hass.async_block_till_done()
-
-    assert result["type"] is FlowResultType.ABORT
-    assert result["reason"] == "value_changed"
-    state = hass.states.get("device_tracker.options_tracker")
-    assert state is not None
-    assert state.attributes[ATTR_LATITUDE] == 41.5
-    assert state.attributes[ATTR_LONGITUDE] == -74.5
-    assert state.attributes[ATTR_GPS_ACCURACY] == 6
-    assert state.attributes[ATTR_BATTERY_LEVEL] == 72
-    assert state.attributes["source"] == "tracker-options"
-
-
-async def test_device_options_flow_updates_registry_metadata(
-    hass: HomeAssistant,
-    config_entry_factory: ConfigEntryFactory,
-) -> None:
-    """Persist device options and update the existing registry record.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-        config_entry_factory: Factory that creates the device config entry.
-    """
-    entry = config_entry_factory(
-        {
-            CONF_ENTITY_PLATFORM: CONF_DEVICE,
-            CONF_NAME: "Options Hub",
-            CONF_YAML_VARIABLE: False,
-            ATTR_MANUFACTURER: "Original",
-            ATTR_MODEL: "Original Model",
-            ATTR_CONFIGURATION_URL: "https://example.com/original",
-        }
-    )
-    assert await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-
-    result = await hass.config_entries.options.async_init(entry.entry_id)
-    assert result["type"] is FlowResultType.FORM
-    assert result["step_id"] == "device_options"
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={
-            ATTR_MANUFACTURER: "Updated",
-            ATTR_MODEL: "Updated Model",
-            ATTR_CONFIGURATION_URL: "https://example.com/updated",
-        },
-    )
-    await hass.async_block_till_done()
-
-    assert result["type"] is FlowResultType.CREATE_ENTRY
-    assert entry.data[ATTR_MANUFACTURER] == "Updated"
-    assert entry.data[ATTR_MODEL] == "Updated Model"
-    device = dr.async_get(hass).async_get_device(identifiers={(DOMAIN, entry.entry_id)})
-    assert device is not None
-    assert device.manufacturer == "Updated"
-    assert device.model == "Updated Model"
-    assert str(device.configuration_url) == "https://example.com/updated"
-
-
-async def test_device_options_invalid_url_preserves_entry_and_registry(
-    hass: HomeAssistant,
-    config_entry_factory: ConfigEntryFactory,
-) -> None:
-    """Keep device data unchanged when its options URL is invalid.
-
-    Args:
-        hass: Home Assistant instance that hosts the integration.
-        config_entry_factory: Factory that creates the device config entry.
-    """
-    entry = config_entry_factory(
-        {
-            CONF_ENTITY_PLATFORM: CONF_DEVICE,
-            CONF_NAME: "Preserved Hub",
-            CONF_YAML_VARIABLE: False,
-            ATTR_MANUFACTURER: "Original",
-            ATTR_MODEL: "Original Model",
-            ATTR_CONFIGURATION_URL: "https://example.com/original",
-        }
-    )
-    assert await hass.config_entries.async_setup(entry.entry_id)
-    await hass.async_block_till_done()
-    registry = dr.async_get(hass)
-    original_device = registry.async_get_device(identifiers={(DOMAIN, entry.entry_id)})
-    assert original_device is not None
-    original_data = dict(entry.data)
-
-    result = await hass.config_entries.options.async_init(entry.entry_id)
-    assert result["type"] is FlowResultType.FORM
-    result = await hass.config_entries.options.async_configure(
-        result["flow_id"],
-        user_input={
-            ATTR_MANUFACTURER: "Should Not Persist",
-            ATTR_MODEL: "Should Not Persist",
-            ATTR_CONFIGURATION_URL: "not a URL",
-        },
-    )
-    await hass.async_block_till_done()
-
-    assert result["type"] is FlowResultType.FORM
-    assert result["step_id"] == "device_options"
-    assert result["errors"] == {"base": "invalid_url"}
-    assert dict(entry.data) == original_data
-    current_device = registry.async_get_device(identifiers={(DOMAIN, entry.entry_id)})
-    assert current_device is not None
-    assert current_device.id == original_device.id
-    assert current_device.manufacturer == "Original"
-    assert current_device.model == "Original Model"
-    assert str(current_device.configuration_url) == "https://example.com/original"
